added sweet alert for deletion item in myList

+++ client/src/pages/MyList.jsx
@@ -1,15 +1,17 @@
 import { FcViewDetails } from "react-icons/fc";
 import { CiEdit } from "react-icons/ci";
 import { MdDeleteOutline } from "react-icons/md";
-import { useEffect, useState } from "react";
+import { useContext, useEffect, useState } from "react";
 import { Link } from "react-router-dom";
+import { AuthContext } from "../providers/Contexts";
+import Swal from "sweetalert2";
 
 const MyList = () => {
-
+    const { user } = useContext(AuthContext)
     const [list, SetList] = useState([])
 
     useEffect(() => {
-        fetch(`https://travel-nest-sigma.vercel.app/spots/test@mail.com`)
+        fetch(`https://travel-nest-sigma.vercel.app/spots/${user?.email}`)
             .then(res => res.json())
             .then(data => {
                 SetList(data)
@@ -18,16 +20,34 @@ const MyList = () => {
     }, [])
 
     const handleDelete = (id) => {
-        fetch(`https://travel-nest-sigma.vercel.app/spots/${id}`, {
-            method: 'DELETE',
-        })
-            .then(res => res.json())
-            .then(data => {
-                console.log(data)
-                if (data.deletedCount > 0) {
-                    SetList(list.filter(item => item._id !== id))
-                }
-            })
+        Swal.fire({
+            title: "Are you sure?",
+            text: "You won't be able to revert this!",
+            icon: "warning",
+            showCancelButton: true,
+            confirmButtonColor: "#d33",
+            cancelButtonColor: "#3085d6",
+            confirmButtonText: "Yes, delete it!"
+        }).then((result) => {
+            if (result.isConfirmed) {
+                fetch(`https://travel-nest-sigma.vercel.app/spots/${id}`, {
+                    method: 'DELETE',
+                })
+                    .then(res => res.json())
+                    .then(data => {
+                        console.log(data)
+                        if (data.deletedCount > 0) {
+                            Swal.fire({
+                                title: "Deleted!",
+                                text: "Your file has been deleted.",
+                                icon: "success",
+                                confirmButtonColor: "#3085d6",
+                            });
+                            SetList(list.filter(item => item._id !== id))
+                        }
+                    })
+            }
+        });
     }
 
     return (
@@ -60,17 +80,18 @@ const MyList = () => {
                             <td><button className="btn btn-outline btn-error text-2xl"><MdDeleteOutline /></button></td>
                         </tr> */}
                         {
-                            list.length > 0 ? list.map((item, idx) =>
-                                <tr key={idx}>
-                                    <td>{idx + 1}</td>
-                                    <td>{item.spot_name}</td>
-                                    <td>{item.location}</td>
-                                    <td>{item.country}</td>
-                                    <td>{item.visitor}</td>
-                                    <td><Link to={`/details/${item._id}`} className="btn btn-outline btn-info text-2xl"><FcViewDetails /></Link></td>
-                                    <td><Link to={`/edit-spot/${item._id}`} className="btn btn-outline btn-warning text-2xl"><CiEdit /></Link></td>
-                                    <td><button onClick={() => handleDelete(item._id)} className="btn btn-outline btn-error text-2xl"><MdDeleteOutline /></button></td>
-                                </tr>
+                            list.length > 0 ? list.map(
+                                (item, idx) =>
+                                    <tr key={idx}>
+                                        <td>{idx + 1}</td>
+                                        <td>{item.spot_name}</td>
+                                        <td>{item.location}</td>
+                                        <td>{item.country}</td>
+                                        <td>{item.visitor}</td>
+                                        <td><Link to={`/details/${item._id}`} className="btn btn-outline btn-info text-2xl"><FcViewDetails /></Link></td>
+                                        <td><Link to={`/edit-spot/${item._id}`} className="btn btn-outline btn-warning text-2xl"><CiEdit /></Link></td>
+                                        <td><button onClick={() => handleDelete(item._id)} className="btn btn-outline btn-error text-2xl"><MdDeleteOutline /></button></td>
+                                    </tr>
                             )
                                 :
                                 <tr>
